create route and controller files
- this call 'model view controller'
- routes = route(URL or endpoint)]
- controller = route handler
- app = setup middleware(route is middleware too)
- receiving req in app.js depending on route the execute one of controller then res sen to req-res cycle
